refactor: move AuthorizedStages helper to pkg/api
Per review feedback, the AuthorizedStages parser does not belong in the
api/v1alpha1 types package (which otherwise contains only methods on
generated types). Move it alongside the other Stage annotation helpers in
pkg/api/stage.go and update callers.
