+++ api/v1alpha1/annotations.go
@@ -34,8 +34,8 @@ const (
 	// resource to indicate that one or more Stages are authorized to manage it.
 	// The value of the annotation is a comma-separated list of
 	// "<project>:<stage>" entries. A single "<project>:<stage>" value is also
-	// valid and authorizes exactly one Stage. Use AuthorizedStages to parse the
-	// value.
+	// valid and authorizes exactly one Stage. The value can be parsed with the
+	// AuthorizedStages helper in pkg/api.
 	AnnotationKeyAuthorizedStage = "kargo.akuity.io/authorized-stage"
 
 	// AnnotationKeyStage is an annotation key that can be set on a resource to
