+++ pkg/api/authorized_stage_test.go
@@ -1,4 +1,4 @@
-package v1alpha1
+package api
 
 import (
 	"testing"
